requirements.txt
Flask
Flask-Bcrypt
pandas
google-api-python-client>=2.84.0
google-genai
langdetect
python-dotenv
pydantic
pytz
gunicorn
psycopg2-binary
matplotlib